Refactor worker templates and configuration
